kinds: five counts stop wearing float — and the validator learns to see all its tables
The render side's carrier split made a question askable that a mixed container
cannot pose — does each param ride the carrier its KIND requires? — and
PAGE_ROWS, PAGE_COLS, PAGES_WIDE, SCROLL_ROWS, WRAP_WIDTH failed it. They were
declared 'measure' because the schema section holding them was literally NAMED
measures: the container defined the kind, the exact inversion the tier split
exists to prevent. Corrected in both tiers (a44a489 carried the schema; this
commit carries the engine).

THE HOLE, exactly as floats-to-ints diagnosed it: allBufs existed and was used
once, while five validation loops held slotBufs — wrong variable in five
places, not a missing concept. And closing the carrier loops alone would NOT
have caught the five (kind 'measure' passes the carrier rule by construction) —
the trap their message named. So this lands three rules together: every loop
that should see every table now does (two stay slotBufs BY REASON, and say so);
a section with no declared carrier is an ERROR (itemBounds sat exactly there,
TOTAL_ROWS riding an undeclared f64 — now declared, justified, and tripwired);
and a lane realizing a semantic param must MATCH its kind, so the tiers can
never disagree silently again. Five validator mutations, five named causes.

The engine: Item's five fields are Int. Truncation happens ONCE, at the
boundary (fixture load), instead of at 37 read sites across seven files —
trunc_nonneg is boundary-only, trunc_nn32 is deleted. On device the item table
splits by carrier (itemMeasures f32x9 + itemExact u32x6, same 9+6 shape the
render side landed independently): page geometry is read native, and the
has_page test is a real != 0 instead of a float > 0.5 proxy.

Executed, not inferred, per their warning about headless blindness: the gpu
suites run real device dispatches through the new table, and a +1 corruption of
the host's IE_PAGE_ROWS lane fails gpu_paginate with page-shaped errors at the
page boundaries — the device provably consumes the u32 table. Reverted.

Fifteen suites green, corpus untouched (fixtures carry f64 VALUES; integer
params are exact through the frozen format — third container migration this
week to ride under it without a byte of drift).

+++ engine/bench/bench.mojo
@@ -74,7 +74,7 @@ def load_input() raises -> BenchInput:
     return BenchInput(bytes^, Trie(block_index^, blocks_m^, blocks_c^))
 
 
-def one_item(byte_count: Int, wrap: Float64) -> List[Item]:
+def one_item(byte_count: Int, wrap: Int) -> List[Item]:
     var it = Item()
     it.byte_count = byte_count
     it.wrap_width = wrap
